fix: scheduling tests failing due to app not being set up correctly in pytest fixture.
